Add files via upload
California Housing Price System (California), Development, and I worked in the data science field. So, we decided to go with Python and got used to the various modules available on the market. We used various datasets to get the various features on which the pricing of the house depends.
Also, by carefully examining the data, we got an idea of what the most valuable features are in deciding the price of the house. Like geographic location, crime rate or nearby facilities, After the model is built and ready to be hosted on the local network, we will add filters to tweak the model or filter out any unwanted results.]